Change the way the scrollbars are displayed
In the gtk2 activity, we hidden the the vertical scrollbar in the webview
setting the policy to NEVER, and showed another independent scollbar,
and moved it doing the calcule as if all the different documents in the epub book
was a single big document.
In gtk3 this is not working, the values needed to do the calculation
are all zero, if the scrollbar si hidden.
The workaround implemented is use a overlay container to place the independent
scrollbar over the unused scrollbar.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>