Save and restore the page with epub files - SL #3413
When the user closed and reopened the activity, reading a epub file,
the page was lost.
The problem had two sources:
The invisible webkit widget used to measure the document height,
can add a scrollbar (randomly) then the inner width change,
the text folw differently and the inner height change. This problem is solved
setting document.documentElement.style.overflow = hidden in the widget
when is used to measure.
The second problem was, after the document was loaded and positioned
the epubview tried to change the scroll position again.
This patch include a minor change in readactivity.py, only to avoid
call many times view.get_current_page.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>