Improve bookmarks add/remove - SL #3405
The bug have two sources.
First, the button used to add/remove the bookmark is a toggle,
and the gtk3 theme is not setting the right grey color, then is not obvious
to the user if the button is pressed or not.
(this problem will be addressed in sugar-artwork)
The second problem is, if the user cancel the dialog creating the bookmark,
the bookmark is not created but the button continue pressed until the page is changed.
This was solved sending a signal to uptade the button state.
Two other problems were found and solved:
* The page number in the bookmark description was wrong by one.
* If the page had more than one bookmark, the clear method only removed the last
then the sidebar accumulated bookmark icons.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>