New design to bookmark visualization
The sidebar where we displayed the bookmark was a thin area
at the left of the screen, was not very intuitive, and was dificult
to access with a finger.
With Gtk3 we can use overlays, then this patch move the sidebar
to the top/right of the screen, and display it with a shape similar
to other book readers, using the user colors.
I have tried to do the botom triangle transparent,
but without success, we can try it later.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>