Show the table of contents in a panel
The table of contents, actually implemented in epub backend,
is displayed in a later panel, instead of in a combo,
to do easier the interaction with touch, and provide context.
The code is now more general to allow the implementation
in the pdf backend too.
The button used to display the table of contents, was moved
from the main toolbar to the view subtoolbar.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>