Fix index management in epub files - SL #3853
Epub index can point to internal links in the content files.
This patch manage this cases. A particular case was found in a
file generated in UY (see attachment in the ticket) where the links
don't point to the start of a file, but to the end of the previous file.
Then, if the index is at the end of a file, we open the next file.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>