Prevent open a file two times
With latests changes (empty panel addition) the canvas is set more than
one time, then the read_file can be triggered more than one time.
Check if a file is already openened to not open it two times.
This prevent a bug where the book is displayed two times, side by side.

This patch also avoid a situation where the book can't be shared because
the self._tempfile is not set when the file is opened at start because
is reopened the activity from the journal.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>