Set the preview in the linkbutton if is not available
When we share bookmarks, the preview images used in the linkbuttons
displayed in the tray, are not saved in the metadata, then are not
received by who receive the book. When the user click in the button
to go to the page the preview is created.
This implementation use a timer to create the image, but should be better
use a event triggered when the viewer has changed to the page.