Do not download again a shared book if is already in the journal
This patch chnge the order of checks, to know if the file is in the journal,
or if is shared, or if is a new instance. Before the first check was if the
file was shared, then even when a file was in the Journal,
tried to download it again.